# Speed up group coordinator tests