Master issue: #830

### Motivation

The `GroupCoordinatorTest` spend too much time to start mock pulsar and shutdown mock pulsar in each tests. See: 
https://github.com/streamnative/kop/issues/830#issuecomment-949259927

### Modifications

1. Each time we start new test method, just restart `GroupCoordinator`, instead of restart new mock pulsar.
2. Remove unused consumer

This PR speed up the `GroupCoordinatorTest`, reduce test time from **500s** down to **14s**

**Before** : 
```text
[INFO] Running io.streamnative.pulsar.handlers.kop.coordinator.group.GroupCoordinatorTest
[INFO] Tests run: 63, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 499.022 s - in io.streamnative.pulsar.handlers.kop.coordinator.group.GroupCoordinatorTest
```
**After** : 
```text
[INFO] Running io.streamnative.pulsar.handlers.kop.coordinator.group.GroupCoordinatorTest
[INFO] Tests run: 63, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 14.275 s - in io.streamnative.pulsar.handlers.kop.coordinator.group.GroupCoordinatorTest
```